Added sign sheet

+++ src/chapter2/exercise3.py
@@ -1,5 +1,5 @@
 Hours = input('Enter hours: ')
-Hours = float(Hours)
+Hours = int(Hours)
 
 Rate = input('Enter rate: ')
 Rate = float(Rate)
